==> notebooks/00_scratchpad.ipynb <==
#### Solo para obtener el formato del json 

import numpy as np
import pandas as pd

# Cargar el dataset
dataset = pd.read_csv('../data/raw/stores_sales_forecasting_updated_v3.1.csv', 
                 sep=';',
                 encoding='utf-8')

# Inspección visual
dataset.head(5)

dataset.iloc[0:10, :].to_json("../data/raw/stores_sales_forecasting_updated_v3.1.json", orient="records")
